requirements.txt
pydicom
Tkinter
Jinja2
